QuickMessage: Auto display soft keyboard on quick reply and other tweaks
This commit adds the automatic display the soft keyboard if the user selects
the 'Quick reply' option from the notification, eliminating an extra click.

It also includes a slight tweak to the strings and re-ordering of the
preferences to make the UX less confusing for end users.

Change-Id: Ic3b57ee80e91f7ba7b1561177cd4a6ea2873f716
Screenshot: https://dl.dropbox.com/u/50993944/Screenshot_2012-11-10-14-09-42.png

Conflicts:

	src/com/android/mms/transaction/MessagingNotification.java